Move more cache functions over
